+++ NEWS.md
@@ -78,4 +78,4 @@
 
 * Updated `fc_export()` to include vector formats (svg, pdf) and to use `ragg` in place of `grDevices` for relevant bitmap formats (png, jpeg, tiff) for improved performance and image quality when `ragg` is installed (#16; @kenkomodo)
 
-* Methods for S3 class `fc` well specified
+* Methods for S3 class `fc` well defined